Remove mongoose-intl
This NPM package was too complicated for the amount of i18n we need to do server-side, and it's not widely used enough to guarantee support, so I took it out. Could still use a simple management system for translations, but we really only need i18n for the event descriptions, so specifying languages manually should work for now.